social_media/requirements.txt
apturl==0.5.2
blinker==1.4
Brlapi==0.8.3
certifi==2020.6.20
chardet==4.0.0
click==8.0.3
colorama==0.4.4
command-not-found==0.3
cryptography~=3.4.7
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
distro==1.7.0
distro-info===1.1build1
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
jeepney==0.7.1
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
louis==3.20.0
macaroonbakery==1.3.1
more-itertools==8.10.0
netifaces==0.11.0
oauthlib==3.2.0
olefile==0.46
pexpect==4.8.0
Pillow==9.0.1
protobuf==3.12.4
ptyprocess==0.7.0
pycairo==1.20.1
pycups==2.0.1
PyGObject==3.42.1
PyJWT==2.4.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
python-apt==2.3.0+ubuntu2.1
python-dateutil~=2.8.2
python-debian===0.1.43ubuntu1
pytz~=2021.1
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
requests~=2.26.0
SecretStorage==3.3.1
six==1.16.0
systemd-python==234
ubuntu-advantage-tools==27.9
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
wadllib==1.3.6
xdg==5
xkit==0.0.0
zipp==1.0.0
Django~=3.2.6
djangorestframework~=3.14.0
cffi~=1.15.1
future~=0.18.2
pip~=22.0.2
attrs~=22.1.0
wheel~=0.37.1
asgiref~=3.5.0
sqlparse~=0.4.2
mysqlclient~=2.1.1
pycparser~=2.21
setuptools~=59.6.0
bcrypt~=3.2.0
djangorestframework-jwt~=1.11.0